pyproject.toml
[project]
name = "transcript-intelligence"
version = "0.1.0"
description = "B2B SaaS call transcript analysis pipeline"
requires-python = ">=3.12"
dependencies = [
    # Pipeline orchestration
    "langgraph>=0.2",
    "langchain-core>=0.3",
    # LLM
    "anthropic>=0.40",
    # Embeddings (local, no API cost)
    "sentence-transformers>=3.0",
    # Clustering
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "scikit-learn>=1.5",
    # Numerical / data
    "numpy>=2.0",
    "pandas>=2.2",
    # Visualizations
    "plotly>=5.20",
    "matplotlib>=3.9",
    "kaleido>=0.2",
    # Slide deck
    "python-pptx>=1.0",
    # Validation & utils
    "pydantic>=2.8",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "rich>=13.0",
    # Notebook
    "ipykernel>=6.29",
    "jupyter>=1.1",
    "nbformat>=5.10",
]

[project.scripts]
transcript-intelligence = "transcript_intelligence.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcript_intelligence"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
